Generates abundances

Generates species abundances based on species response functions and environmental variables.

Usage

abundances(env, spp, nc)

Arguments

env

Environmental variables. Usually generated by make.env. Users may as well supply own environmental variables.

spp

Species parameters. Usually generated by species.

nc

Number of counts to be simulated. If omitted no simulation of the counting process is carried out.

Value

spp

Data frame containing species abundances.

env

Data frame containing environmental variables.

References

Minchin, P.R. (1987) Multidimensional Community Patterns: Towards a Comprehensive Model. Vegetatio, 71, 145-156.

Author

Mathias Trachsel and Richard J Telford

See also

Examples

spec <- species(nspp = 30,ndim = 10,Amax = runif,fun = runif, xpar = c(-50,150),
    srange = 200, alpha = 4, gamma = 4)
env.var <- make.env(100,elen =rep(100,10),emean = rep(50,10), edistr ='uniform',
    ndim = 10)
spec.abun <- abundances(env.var,spec,200)