Significance Tests of Quantitative Palaeoenvironmental Reconstructions

Significance tests for quantitative palaeoenvironmental reconstructions derived from transfer functions. Functions from the autocorTF package are now included in palaeoSig.

Details

This package includes:

significance tests for quantitative palaeoenvironmental reconstructions (randomTF, obs.cor)

graphical methods to show autocorrelation in transfer functions (rne)

null model test of transfer functions performance in a spatially autocorrelated environment - see vignette.

Several functions have from autocorTF version 1.0 and palaeoSig version 1.0 have been rewritten or replaced with more flexible or user friendly functions. See news(package="palaeoSig") for details.

See also my blog at https://quantpalaeo.wordpress.com/

Package:	palaeoSig

Type:	Package

Version:	2.0-7

Date:	2022-11-29

License:	GPL - 3

LazyLoad:	yes

Author

Richard Telford richard.telford@uib.no

References

Telford, R. J. and Birks, H. J. B. (2009) Evaluation of transfer functions in spatially structured environments. Quaternary Science Reviews 28: 1309--1316. doi:10.1016/j.quascirev.2008.12.020

Telford, R. J. and Birks, H. J. B. (2011) A novel method for assessing the statistical significance of quantitative reconstructions inferred from biotic assemblages. Quaternary Science Reviews 30: 1272--1278. doi:10.1016/j.quascirev.2011.03.002

See also